updated fragment

+++ app/src/main/java/com/invest/advisor/data/network/yahooResponse/RegularMarketVolume.kt
@@ -4,7 +4,7 @@ package com.invest.advisor.data.network.yahooResponse
 import com.google.gson.annotations.SerializedName
 
 data class RegularMarketVolume(
-    val raw: Int,
+    val raw: Double,
     val fmt: String,
     val longFmt: String
 )
